+++ src/Sulu/Bundle/SecurityBundle/Entity/UserSettingRepository.php
@@ -12,6 +12,7 @@
 namespace Sulu\Bundle\SecurityBundle\Entity;
 
 use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\NoResultException;
 use Sulu\Component\Security\Authentication\UserSettingRepositoryInterface;
 
 /**
@@ -20,4 +21,27 @@ use Sulu\Component\Security\Authentication\UserSettingRepositoryInterface;
  */
 class UserSettingRepository extends EntityRepository implements UserSettingRepositoryInterface
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function getSettingsByKeyAndValue($key, $value)
+    {
+        if (is_array($value)) {
+            $value = json_encode($value);
+        }
+
+        try {
+            $qb = $this->createQueryBuilder('setting')
+                ->where('setting.key=:key')
+                ->andWhere('setting.value=:value');
+
+            $query = $qb->getQuery();
+            $query->setParameter('key', $key);
+            $query->setParameter('value', $value);
+
+            return $query->getResult();
+        } catch (NoResultException $ex) {
+            return;
+        }
+    }
 }